feat: color range passed from the frontend

+++ colorUtils/grayscale.go
@@ -1,7 +1,7 @@
 package colorUtils
 
-func RGBtoGrayscale(r, g, b uint8) uint8 {
+func RGBtoGrayscale(r, g, b uint8) float64 {
 	// Standard grayscale conversion using luminance formula
-	gray := uint8(0.299*float64(r) + 0.587*float64(g) + 0.114*float64(b))
+	gray := float64(0.299*float64(r) + 0.587*float64(g) + 0.114*float64(b))
 	return gray
 }
